+++ process-manager-api/src/main/java/org/orbisgis/processmanagerapi/inoutput/IInOutPut.java
@@ -129,4 +129,14 @@ public interface IInOutPut {
      * @return The keywords of the input/output.
      */
     String[] getKeywords();
+
+    /**
+     * Methods used by groovy to simplify the syntax.
+     *
+     * @param name Name of the methods.
+     * @param args Object array of the method argument.
+     *
+     * @return Method result.
+     */
+    Object methodMissing(String name, Object args);
 }
